LED Health indicator, fixed health values
Added code to cause the LED to turn green for good health, blue for iffy, and red for bad. Removed code that killed the bot if it slept for less than a minute for demonstration purposes, and replaced with code that would cut health by 75% if it sleeps for less than four hours. Added code that increased health if the thing was fed.